move from `ctypes/__init__.pyi` to `_ctypes.pyi` - `_CField` - `_StructUnionMeta` - `_StructUnionBase` - `Union` - `Structure`
